requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.10
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-yasg==1.20.0
filelock==3.4.0
flower==1.0.0
humanize==3.14.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
packaging==21.3
pipenv==2021.11.23
platformdirs==2.4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.13.3
